fixup!: rebase and fix counts
