+++ kobo/apps/accounts/utils.py
@@ -1,11 +1,20 @@
 from allauth.socialaccount.models import SocialAccount
 from django.conf import settings
+from django.contrib.auth.hashers import UNUSABLE_PASSWORD_PREFIX
+from django.db.models import CharField, Count, F, Func, Q, Value
+from django.db.models.functions import Lower
 
 from kobo.apps.accounts.models import SocialAppManagedDomain
+from kobo.apps.help.models import InAppMessageUsers, MessageType
 from kobo.apps.kobo_auth.shortcuts import User
 from kobo.apps.stripe.constants import ACTIVE_STRIPE_STATUSES
 
 SOCIAL_APP_IDENTIFIER = 'socialaccounts.SocialApp'
+
+
+class SplitPart(Func):
+    function = 'SPLIT_PART'
+    output_field = CharField()
 
 
 def user_has_inactive_paid_subscription(username):
@@ -58,3 +67,36 @@ def user_is_managed_by_sso(user):
         ).exists()
         return user_has_social_account
     return False
+
+
+def users_needing_update(social_app: 'socialaccount.SocialApp', domain: str):
+    users_already_received_message = InAppMessageUsers.objects.filter(
+        in_app_message__message_type=MessageType.MANAGED_SSO_REMINDER,
+        in_app_message__generic_related_objects__contains={
+            SOCIAL_APP_IDENTIFIER: social_app.pk,
+        },
+    ).values_list('user_id', flat=True)
+    users = (
+        User.objects.prefetch_related('socialaccount_set')
+        .filter(extra_details__sso_exempt=False)
+        .exclude(id__in=users_already_received_message)
+        .annotate(
+            domain=SplitPart(Lower(F('email')), Value('@'), Value(2)),
+            managed_account=Count(
+                'socialaccount',
+                filter=Q(socialaccount__provider=social_app.provider_id),
+            ),
+            # TODO: why doesn't exclude work here?
+            other_accounts=Count(
+                'socialaccount',
+                filter=~Q(socialaccount__provider=social_app.provider_id),
+            ),
+        )
+        .filter(domain=domain)
+        .exclude(
+            password__startswith=UNUSABLE_PASSWORD_PREFIX,
+            other_accounts=0,
+            managed_account=1,
+        )
+    )
+    return users
